requirements.txt
amqp==2.4.2
asn1crypto==0.24.0
bcrypt==3.1.6
billiard==3.5.0.5
celery==4.2.2
certifi==2019.3.9
cffi==1.12.2
cftime==1.0.3.4
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.4.2
Django==2.1.7
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
kombu==4.3.0
MarkupSafe==1.1.1
netCDF4==1.5.0.1
numpy==1.16.2
openapi-codec==1.3.2
paramiko==2.4.2
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pysftp==0.2.9
pytz==2018.9
PyYAML==5.1
requests==2.21.0
scp==0.13.2
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
vine==1.3.0
whitenoise==4.1.2